+++ Src/Library/changelog.md
@@ -1,2 +1,6 @@
+### CHANGES
+- `AddAuthenticationJWTBearer()` has been renamed to `AddJWTBearerAuth()` and signature changed
+
 ### NEW
 - customize dto property binding failure message #356
+- ability to specify jwt bearer events with `AddJWTBearerAuth()` #359